Make get_atomic_grid return indices
The method now returns a tuple of two elements. The first one is the
atomic or local grid that is expected. The second one are the indices
of the points (from the molecular grid) that constitute the atomic grid.
These will be used to keep track of the function values to use for the
numerical integration (in the atomic grid).