Remove points and weights args from AngularGrid initialization
If the user has a set of points and corresponding weights for making an angular grid,
they can just use the Grid class. I don't see any point in allowing the user pass it
as arguments to AngularGrid alongside a degree/size that many not match the points/weights.
In addition, this functionality was not even used or tested. Supporting this, we need to add
further tests and checks to prevent misuse, but most importantly, I don't believe it makes
sense from a design principle perspective.
The points and weights setters were added to make it possible to modify AngularGrid attributes
without needing to make a new instance.